+++ lib/vey-daemon/src/listen/udp/listen.rs
@@ -539,7 +539,7 @@ where
         if listen_config.use_ebpf(rustix::process::getuid().as_raw()) {
             match UdpSocketSelector::new(
                 rustix::process::getpid().as_raw_pid(),
-                self.server.version() as u32,
+                self.server.version() as u16,
                 listen_config.address(),
                 self.conn_track.ebpf_conn_track_size(),
             ) {
